Change i32 type -> byte
This change affects the Self schema for XAMS.

+++ doc/schemas/Thrift/self.thrift
@@ -5,5 +5,5 @@ struct Device {
 
 struct _Self {
   1: required Device device;
-  2: optional i32 direction;
+  2: optional byte direction;
 }
